chore: add more unit tests
